gateway/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-snap7==1.3
paho-mqtt==2.0.0
pydantic==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
websockets==12.0
